sort the from_server_role nameservers to avoid over-eager rewriting
If we don't do this, we create the resolv.conf with the nameservers
in the order found by search.  That's unpredictable ordering, which
means that the resolv.conf file is re-generated on almost every run
of chef-client. That means lots of cruft in backup files, at least.

The commit message has been fully justified to ease in its reading.